+++ src/solutions/p005.rs
@@ -2,7 +2,7 @@ use super::Solve;
 use crate::utils::math::lcm;
 
 pub struct Parameters {
-    pub n: u64,
+    pub n: usize,
 }
 
 impl Solve for Parameters {
