+++ src/utils/fibonacci.rs
@@ -1,14 +1,14 @@
-pub struct Fibonacci(u64, u64);
+pub struct Fibonacci(usize, usize);
 
 impl Fibonacci {
-    pub fn new(s1: u64, s2: u64) -> Fibonacci {
+    pub fn new(s1: usize, s2: usize) -> Fibonacci {
         Fibonacci(s1, s2)
     }
 }
 
 impl Iterator for Fibonacci {
-    type Item = u64;
-    fn next(&mut self) -> Option<u64> {
+    type Item = usize;
+    fn next(&mut self) -> Option<usize> {
         let next = self.0 + self.1;
         self.0 = self.1;
         self.1 = next;
